Don't hide numba behind an abstraction
Originally, I created drop-in replacements for numba.njit and
numba.prange that would work whether or not numba was available.

As it turns out, this leads to errors at runtime because Numba looks
for information inside symbols to try to deduce types.

The right solution is to change numba_guard so that when Numba is
available, numba_guard.prange and numba_guard.njit are actually
numba.prange and numba.njit.